+++ src/util.rs
@@ -1,25 +1,6 @@
 use syn;
 
-use syn::parse::Error;
-use syn::spanned::Spanned;
 use quote::ToTokens;
-
-// Returns error if there is more than one.
-pub fn map_only_one<S, T, F>(iter: &[S], dlg: F) -> Result<Option<T>, Error>
-where F: Fn(&S) -> Result<Option<T>, Error>,
-      S: Spanned,
-{
-    let mut result = None;
-    for item in iter {
-        if let Some(answer) = dlg(item)? {
-            if result.is_some() {
-                return Err(Error::new(item.span(), "Multiple #[builder] on the same field"))
-            }
-            result = Some(answer);
-        }
-    }
-    Ok(result)
-}
 
 pub fn path_to_single_string(path: &syn::Path) -> Option<String> {
     if path.leading_colon.is_some() {